(GA) align with main (#101)
* (FRO) Removing data folder from main folder, the only file there was moved to p3/aoSystem/data (#94)

* (GA) Fix telescopeOtf (#98)

* (GA) pupil as int and not bool

* (GA) telescopeOtf when samp < 1

* (GA) new fix for telescopeOtf (#99)

* (GA) telescopeOtf bugfix

* (GA) bilinear interp with zoom

* (GA) np --> nnp and cpuArray in interpolateSupport (#100)

---------

+++ p3/aoSystem/telescope.py
@@ -105,10 +105,10 @@ class telescope:
             th = self.pupilAngle*nnp.pi/180
             x = nnp.linspace(-D/2, D/2, resolution)
             X,Y = nnp.meshgrid(x,x)
-            Xr = X*nnp.cos(th) + Y*nnp.sin(th)
-            Yr = Y*nnp.cos(th) - X*nnp.sin(th)
-            R = nnp.hypot(Xr, Yr)
-            P = (R <= self.R) * (R > self.R*self.obsRatio)
+            Xr  = X*nnp.cos(th) + Y*nnp.sin(th)
+            Yr  = Y*nnp.cos(th) - X*nnp.sin(th)
+            R   = nnp.hypot(Xr,Yr)
+            P   = ((R <= self.R) * (R > self.R * self.obsRatio)).astype(int)
             self.pupil = np.asarray(P)
             self.verb = False
